code for adding delete button and delte user functionality

+++ dom.js
@@ -8,32 +8,41 @@ function test(event)
     var tm = document.getElementById('time').value;
 
     let myObj = {name,em,ph,dt,tm};
-    localStorage.clear();
+    
 
     let myObj_serialized = JSON.stringify(myObj)
     console.log(myObj_serialized)
 
     localStorage.setItem('myObj',myObj_serialized)
    
+    showOnScreen(myObj);
+}
+function showOnScreen(myObj)
+{
     let myObj_dserialized = JSON.parse(localStorage.getItem('myObj'));
     var nm = myObj_dserialized.name
     var a = myObj_dserialized.em;
     var b = myObj_dserialized.ph
-    var fm = document.getElementById('form');
-    var para = document.createElement('pre');
-    para.style.fontSize = '20 px'
-    para.style.fontWeight = 'bold'
-    para.textContent = "Name = " + nm + " Email = " + a + " Phone = " + b;
-    fm.append(para)
-
-   
-
-   
-
-    
-
-    
-
+    var ul = document.getElementById('list');
+    var li = document.createElement('li');
+    // para.style.fontSize = '20 px'
+    // para.style.fontWeight = 'bold'
+    li.textContent = "Name = " + nm + " Email = " + a + " Phone = " + b;
+    li.style.fontWeight = 'bold'
+    ul.append(li);
+    var btn = document.createElement('button');
+    btn.className = 'btn btn-primary delete'
+    btn.textContent = 'DELETE'
+    btn.addEventListener("click", function(e){
+        if(confirm('are you sure'))
+                {
+                    localStorage.clear();
+                    var li = e.target.parentElement;
+                    var item = document.getElementById('list');
+                    item.removeChild(li);
+                }
+    });
+    li.append(" ",btn)
 
 }
